refactor: 당첨 순위 찾는 기능을 WinningRank에서 Lottos로 이동
