+++ src/main/kotlin/lotto/model/Lottos.kt
@@ -4,4 +4,6 @@ class Lottos(
     val lottos: List<Lotto>
 ) {
     val size = lottos.size
+
+    fun findRanks(winningNumbers: Lotto) = lottos.map { it.findRank(winningNumbers) }
 }
