+++ src/main/kotlin/lotto/model/WinningRank.kt
@@ -10,14 +10,4 @@ enum class WinningRank(
     THIRD_PRIZE(4, 50_000),
     FOURTH_PRIZE(3, 5_000),
     NONE(0, 0);
-
-    companion object {
-        fun findRanks(lottos: Lottos, winningNumbers: Lotto) = lottos.lottos.map { findRank(it, winningNumbers) }
-
-        private fun findRank(lotto: Lotto, winningNumbers: Lotto): WinningRank {
-            return values().find {
-                it.matchedNumberCount == lotto.findMatchedNumberCount(winningNumbers)
-            } ?: NONE
-        }
-    }
 }
